requirements.txt
aiogram>=3.0.0
numpy>=1.20.0
faiss-cpu>=1.7.0
python-dotenv>=0.19.0
motor>=3.0.0
pymongo>=4.0.0